+++ src/model/movie-model.js
@@ -1,7 +1,7 @@
 import {generateMovies} from '../mock/movie';
 import {generateComments} from '../mock/comments';
 
-const MOVIES_COUNT = 7;
+const MOVIES_COUNT = 70;
 const TOTAL_COMMENTS_COUNT = 40;
 
 export default class MovieModel {
