+++ src/presenter/movie-presenter.js
@@ -9,8 +9,8 @@ import MovieContainerView from '../view/movie-container-view';
 import MovieListView from '../view/movie-list-view';
 import ShowMoreButtonView from '../view/show-more-button-view';
 import MovieCardView from '../view/movie-card-view';
-import MovieDetailsView from '../view/movie-popup-view';
-import MovieDetailsCommentView from '../view/movie-popup-comment-view';
+
+import MovieDetailsPresenter from '../presenter/movie-details-presenter';
 
 const ALL_MOVIE_COUNT_PER_STEP = 5;
 const TOP_RATED_MOVIE_COUNT_PER_STEP = 2;
@@ -40,16 +40,14 @@ export default class MainPresenter {
   #mostCommentedMovieListComponent = new MovieListView(SectionTitle.mostCommented, true);
   #mostCommentedMovieListContainerComponent = new MovieContainerView();
   #loadMoreButtonComponent = new ShowMoreButtonView();
-  #movieDetailsComponent = null;
 
-  constructor (mainContainer, profileElement, footerStatisticsElement, movieModel) {
+  constructor(mainContainer, profileElement, footerStatisticsElement, movieModel) {
     this.#mainContainer = mainContainer;
     this.#profileElement = profileElement;
     this.#footerStatisticsElement = footerStatisticsElement;
     this.#movieModel = movieModel;
     this.#comments = [...this.#movieModel.comments];
     this.#movies = [...this.#movieModel.movies];
-    // this.#movies = [...movies];
   }
 
   init = () => {
@@ -88,38 +86,12 @@ export default class MainPresenter {
     this.#movies.sort((a, b) => b.comments.length - a.comments.length).slice(0, Math.min(this.#movies.length, MOST_COMMENTED_MOVIE_COUNT_PER_STEP)).forEach((movie) => this.#renderMovie(movie, this.#mostCommentedMovieListContainerComponent.element));
   }
 
-  #renderMovie (movie, MovieCountainerElement) {
+  #renderMovie(movie, MovieCountainerElement) {
     const movieCardComponent = new MovieCardView(movie);
     render(movieCardComponent, MovieCountainerElement);
 
-    const onEscKeyDown = (evt) => {
-      if (evt.key === 'Escape' || evt.key === 'Esc') {
-        closeMovieDetails(evt, this.#movieDetailsComponent);
-      }
-    };
-
-    const handleCloseButtonClick = (evt) => {
-      closeMovieDetails(evt, this.#movieDetailsComponent);
-    };
-
-    function closeMovieDetails(evt, component) {
-      evt.preventDefault();
-      component.element.remove();
-      component.removeElement();
-      document.removeEventListener('keydown', onEscKeyDown);
-      document.body.classList.remove('hide-overflow');
-    }
-
-    const onCardClick = () => {
-      document.body.classList.add('hide-overflow');
-      this.#movieDetailsComponent = new MovieDetailsView(movie);
-      render(this.#movieDetailsComponent, document.body);
-      this.#comments.filter((comment) => movie.comments.includes(comment.id)).forEach((comment) => render(new MovieDetailsCommentView(comment), this.#movieDetailsComponent.element.querySelector('.film-details__comments-list')));
-      document.addEventListener('keydown', onEscKeyDown);
-      this.#movieDetailsComponent.element.querySelector('.film-details__close-btn').addEventListener('click', handleCloseButtonClick);
-    };
-
-    movieCardComponent.element.querySelector('.film-card__link').addEventListener('click', onCardClick);
+    const movieDetailsPresenter = new MovieDetailsPresenter(movieCardComponent, movie, this.#comments);
+    movieDetailsPresenter.init();
   }
 
   #handleLoadMoreButtonClick = (evt) => {
